+++ PSFramework/internal/tepp/scripts/license-name.ps1
@@ -1,0 +1,3 @@
+﻿Register-PSFTeppScriptblock -Name 'PSFramework-license-name' -ScriptBlock {
+	(Get-PSFLicense).Product
+}